pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-orchestrator"
version = "0.1.0"
description = "Production-minded local CLI for autonomous repo orchestration"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Open Source Contributors" }]
dependencies = [
  "typer>=0.12.0",
  "SQLAlchemy>=2.0.0",
  "PyYAML>=6.0",
  "platformdirs>=4.2.0",
  "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.5.0",
]

[project.scripts]
execforge = "orchestrator.cli.main:main"
agent-orchestrator = "orchestrator.cli.main:main"
orchestrator = "orchestrator.cli.main:main"
agent-controlplane = "orchestrator.cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
